chore: #17 바텀 모달 이름 변경
