+++ app/_layout.tsx
@@ -14,7 +14,6 @@ import { useColorScheme, Platform } from 'react-native';
 import { StatusBar } from 'expo-status-bar';
 import { SafeAreaProvider, SafeAreaView } from 'react-native-safe-area-context';
 import Toast from '@/components/Toast';
-import { BottomSheetModalProvider } from '@gorhom/bottom-sheet';
 
 export {
   // Catch any errors thrown by the Layout component.
@@ -60,37 +59,28 @@ function RootLayoutNav() {
   const colorScheme = useColorScheme();
 
   return (
-    <GestureHandlerRootView>
-      <BottomSheetModalProvider>
-        <SafeAreaProvider>
-          {/* StatusBar 설정 */}
-          <StatusBar
-            style={colorScheme === 'dark' ? 'light' : 'dark'}
-            backgroundColor="#F0ECFE"
-            translucent={Platform.OS === 'ios'}
-          />
+    <SafeAreaProvider>
+      {/* StatusBar 설정 */}
+      <StatusBar
+        style={colorScheme === 'dark' ? 'light' : 'dark'}
+        backgroundColor="#F0ECFE"
+        translucent={Platform.OS === 'ios'}
+      />
 
-          {/* iOS 대응: SafeAreaView top 영역 배경 적용 */}
-          <SafeAreaView
-            edges={['top']}
-            style={{ backgroundColor: '#F0ECFE' }}
-          />
+      {/* iOS 대응: SafeAreaView top 영역 배경 적용 */}
+      <SafeAreaView edges={['top']} style={{ backgroundColor: '#F0ECFE' }} />
 
-          {/* 앱 전체 Theme 적용 */}
-          <ThemeProvider
-            value={colorScheme === 'dark' ? DarkTheme : DefaultTheme}
-          >
-            <Stack
-              screenOptions={{
-                headerShown: false,
-              }}
-            >
-              <Stack.Screen name="modal" options={{ presentation: 'modal' }} />
-            </Stack>
-            <Toast />
-          </ThemeProvider>
-        </SafeAreaProvider>
-      </BottomSheetModalProvider>
-    </GestureHandlerRootView>
+      {/* 앱 전체 Theme 적용 */}
+      <ThemeProvider value={colorScheme === 'dark' ? DarkTheme : DefaultTheme}>
+        <Stack
+          screenOptions={{
+            headerShown: false,
+          }}
+        >
+          <Stack.Screen name="modal" options={{ presentation: 'modal' }} />
+        </Stack>
+        <Toast />
+      </ThemeProvider>
+    </SafeAreaProvider>
   );
 }
